Add seeds for user assessment claims
Updates mileage to a decimal
AB#10656
NOTE: Required a DB update

+++ Fmas12d.Api/ViewModels/UserAssessmentClaim.cs
@@ -29,7 +29,7 @@ namespace Fmas12d.Api.ViewModels
     public DateTimeOffset? ExportedDate { get; set; }
     public int Id { get; set; }
     public DateTimeOffset LastUpdated { get; set; }
-    public int? Mileage { get; set; }
+    public decimal? Mileage { get; set; }
     public decimal? MileagePayment { get; set; }
 
     public static Func<Business.Models.UserAssessmentClaim, UserAssessmentClaim> ProjectFromModel
